models/requirements.txt
# ================= CORE =================
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

tensorflow==2.16.1
keras==3.3.3

# ================= VISIÓN =================
opencv-python==4.9.0.80
mediapipe==0.10.31
Pillow>=10.0.0

# ================= ML =================
scikit-learn>=1.3.0
scipy>=1.11.0
numpy>=1.26.0
pandas>=2.0.0

# ================= VISUALIZACIÓN (MINIMAL) =================
matplotlib>=3.7.0

# ================= UTILIDADES =================
pyyaml>=6.0.0
tqdm>=4.65.0
rich>=13.0.0
psutil>=5.9.0
